Se intenta rehacer el tema de los pipes

+++ cliente.c
@@ -21,16 +21,27 @@ int main(int argc, char *argv[]) {
     // Abrir el pipe como escritura
     printf("ID: %d, pipeNom: %s\n", args.id, args.pipeNom);
     int fd = open(args.pipeNom, O_WRONLY);
-    // Abrir un pipe de nombre IDPIPE para la comunicación exclusiva entre el Gestor y el Cliente
-    char idPipe[100];
-    sprintf(idPipe, "%dPIPE", args.id);
-    mkfifo(idPipe, 0666);
+    // testear si se abrio correctamente
+    if (fd == -1){
+        perror("Error al abrir el pipe");
+        exit(EXIT_FAILURE);
+    }
+
     // Enviar el identificador del usuario al Gestor usando el pipe como ID #
     char mensaje[100];
-    sprintf(mensaje, "ID %d", args.id);
-    write(fd, mensaje, 100);
+    sprintf(mensaje, "ID%d", args.id);
+
+    // Abrir un pipe de nombre IDPIPE para la comunicación exclusiva entre el Gestor y el Cliente
+    char idPipe[100];
+    sprintf(idPipe, "pipeCliente%d", args.id);
+    if (mkfifo(idPipe, 0666) == -1) {
+        perror("Error al crear el pipe");
+        exit(EXIT_FAILURE);
+    }
+
     // Cerrar el pipe
-    close(fd);
+    // close(fd);
+
     // Abrir el pipe de nombre IDPIPE como lectura
     unlink(idPipe); // Eliminar el pipe si existe
     fd = open(idPipe, O_RDONLY);
